refactor(diagnostics): split into a package + fix viz_example docs asset (#235)
Split the 1309-line diagnostics.py into a package with _colormaps /
_colors / _fonts submodules (the three groups share no helpers). Public
API and all import paths (dartwork_mpl.diagnostics.*, the top-level
namespace, the deprecated asset_viz alias, the private _plot_single_library
re-export tests use) are unchanged.

Also fix the long-standing intermittent viz_example docs-build failure
(root cause, now reproduced deterministically): _save_viz_example passed
dm.cm(15)/dm.cm(10) Length objects to fig.set_size_inches, so matplotlib
built an object-dtype array and raised 'ufunc isfinite not supported',
the SVG was never written, and the -W build aborted on the missing
figure. Size via dm.figsize('15cm','10cm') (inch floats). As
defense-in-depth, build_api_assets retries each generator once and writes
a visible placeholder SVG on final failure so one bad asset can't abort
the docs build. New regression tests in test_docs_assets.py.
